+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Revamped `cipher` package
+- Added box algorithm to `cipher` package
 - Added regex support to `blacklist`
 
 ## [v0.3.1]